Move mp's ActionBar to ~/components
The idea is to drop all them redundant directory layers. My long term memory is great and yet I can't remember where things are.

Next to address (soonish): containers vs components. Damn son!
